% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mleshsIIreg}
\alias{mleshsIIreg}
\title{ML fit of Linear regression model with skew heperbolic secant type II distributed errors by using \code{optim} function.}
\usage{
mleshsIIreg(y, x = NULL, initial, gr = NULL, method = "BFGS",
  lower = -Inf, upper = Inf, control = list(), hessian = FALSE)
}
\arguments{
\item{y}{response vector.}

\item{x}{design matrix.}

\item{initial}{vector of initial values.}

\item{method}{The method to be used. See \optim for Detailes.}

\item{lower, upper}{Bounds on the variables for the "L-BFGS-B" method, or bounds in which to search for method "Brent".}

\item{control}{A list of control parameters. See \optim for Detailes.}

\item{hessian}{Logical. Should a numerically differentiated Hessian matrix be returned?}
}
\value{
\item{coffitients}{estimation of coffitients.}

\item{sigma}{estimation of sigma.}

\item{alpha}{estimation of alpha}

\item{loglik}{value of log-likelihood function at estimated parameters.}

\item{hessian}{estimated hessian matrix.}

\item{counts}{See \link{optim} for Detailes.}

\item{convergents}{See \link{optim} for Detailes.}
}
\description{
ML fit of Linear regression model with SHS type II distributed errors by using \code{optim} function.
}
\examples{
beta = c(1, 2, 3); sigma = 2; alpha = 0.3; n = 10000;
e    = rshsII(n,mu = 0,sigma = sigma,alpha = alpha)  #A random sample
x1   = rnorm(n); x2 = rnorm(n); #covariates
y    = beta[1] + beta[2] * x1 + beta[3] * x2 + e
regfit = mleshsIIreg(y, x=cbind(rep(1, length(y)), x1, x2),
                    initial = c(beta, sigma, alpha), method = "Nelder-Mead")
cbind(c(beta, sigma, alpha), c(regfit$coefficient, regfit$sigma, regfit$alpha))
}
\author{
Jamil Ownuk <\email{JamilOwnuk@yahoo.com}>, Hossein Baghishani <\email{hbaghishani@yahoo.com}> and Ahmad Nezakati <\email{anezakati591@yahoo.com}>
}

